Move plan check to the page level
